fix: use exclusively Typer CLI
I had a mix between a CLI with Typer (in cli.py) and a main script (in main.py). I did the fusion between these two, and I now have a CLI with 2 commands: import-data and export-md-lesions-to-csv.

I also put this in the README file.
